pyproject.toml
[project]
name = "codspeed-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-codspeed>=3.0.0b4",
]
